Move FindNodeAddress to resource package
